Remove timeouts entirely.
I expect querying a mailbox to return back the state of the mailbox at the time of the request, not to enter a loop which anticipates an alternate state. Moreover, this looping is only done on empty inboxes, so it only meets the goal of guaranteeing a new email has arrived in one use-case. If you have an inbox which already has data in it, you will be immediately returned the results with no certainty that said results contain the new email you are waiting for.

As such, a developer will always need to wrap their inbox lookups in a loop that analyze the resulting emails, checking to see if their new emails has arrived (perhaps looking for a timestamp, a certain subject, or a certain sender). The timeout code gets in the way of this as it does unnecessary extra looping for empty inboxes. 

This updated code is cleaner and leaves the looping to the developer. An example of analyzing email results is below

```ruby
  10.times do |i|
    emails = @mailbox.get_emails_by_recipient(inbox)
    if emails.is_a?(Array) && emails.map(&:subject).include?(subject)
      break
    else
      sleep(2)
      raise "expected email not received for #{inbox}" if i == 9
    end
  end```