Remove Timeouts
Background: I expect querying a mailbox to return back the state of the mailbox at the time of the request. It is undesirable for the library to enter a loop which anticipates an alternate state. Moreover, the looping mechanism only applies to empty inboxes. As such, the goal of guaranteeing a new email has arrived is only valuable to the empty-inbox use-case. If you have an inbox which already has data in it, you will be immediately returned the results with no certainty that said results contain the new email you are waiting for.

A developer will always need to wrap their inbox lookups in a loop that analyze the resulting emails to check if their new email has arrived (perhaps looking for a timestamp, a certain subject, or a certain sender). If this is the case, the timeout code gets in the way of this as it does unnecessary extra looping for empty inboxes. 

This updated code is cleaner and leaves the looping to the developer. An example of analyzing email results is below

```ruby
  10.times do |i|
    emails = @mailbox.get_emails_by_recipient(inbox)
    if emails.is_a?(Array) && emails.map(&:subject).include?(subject)
      break
    else
      sleep(2)
      raise "expected email not received for #{inbox}" if i == 9
    end
  end```